+++ app/templates/_gulpfile.js
@@ -1,8 +1,10 @@
 'use strict';
 
 var gulp = require('gulp');
+var gulpUtil = require('gulp-util');
 var del = require('del');
 var webpack = require('gulp-webpack');
+var webpackConfigs = require('./webpack.config.js');
 var named = require('vinyl-named');
 var minifyHtml = require('gulp-minify-html');
 var uglify = require('gulp-uglify');
@@ -13,15 +15,53 @@ var browserSync = require('browser-sync');
 var modRewrite = require('connect-modrewrite');
 var runSequence = require('run-sequence');
 var karma = require('gulp-karma');
+var opn = require('opn');
+
+var isBuild = false;
+var isOpen = false;
 
 gulp.task('clean', function () {
   del.sync(['.tmp', 'dist']);
 });
 
 gulp.task('webpack', function () {
+  webpackConfigs.quiet = !isBuild;
   return gulp.src(['app/*.{js,jsx}'])
     .pipe(named())
-    .pipe(webpack(require('./webpack.config.js')))
+    .pipe(webpack(webpackConfigs))
+    .pipe(gulp.dest('.tmp'));
+});
+
+gulp.task('webpack:watch', function () {
+  webpackConfigs.watch = true;
+  return gulp.src(['app/*.{js,jsx}'])
+    .pipe(named())
+    .pipe(webpack(webpackConfigs, null, function(error, stats){
+      if(error){
+        console.log(error);
+      }else{
+        gulpUtil.log(stats.toString({
+          colors: gulpUtil.colors.supportsColor,
+          hash: false,
+          timings: false,
+          chunks: false,
+          chunkModules: false,
+          modules: false,
+          children: true,
+          version: true,
+          cached: false,
+          cachedAssets: false,
+          reasons: false,
+          source: false,
+          errorDetails: false
+        }));
+        if(!isOpen){
+          opn('http://localhost:3000', null, function(){
+            isOpen = true;
+          });
+        }
+      }
+    }))
     .pipe(gulp.dest('.tmp'));
 });
 
@@ -63,11 +103,7 @@ gulp.task('copy', function () {
     .pipe(gulp.dest('dist'));
 });
 
-gulp.task('preServe', function (callback) {
-  runSequence('clean', 'webpack', callback);
-});
-
-gulp.task('serve', ['preServe'], function () {
+gulp.task('browserSync', function(callback){
   browserSync({
     files: ['app/*.html', '.tmp/**/*'],
     server: {
@@ -77,10 +113,16 @@ gulp.task('serve', ['preServe'], function () {
           '!\\.\\w+$ /index.html [L]'
         ])
       ]
-    }
+    },
+    port: 3000,
+    reloadOnRestart: false,
+    open: false
   });
+  callback();
+});
 
-  gulp.watch(['app/**/*', '!app/**/*.html'], ['webpack']);
+gulp.task('serve', function (callback) {
+  runSequence('clean', 'webpack', 'browserSync', 'webpack:watch', callback);
 });
 
 gulp.task('test', function () {
@@ -92,6 +134,7 @@ gulp.task('test', function () {
 });
 
 gulp.task('build', function (callback) {
+  isBuild = true;
   runSequence('clean', 'html', 'scripts', 'styles', 'images', 'fonts', 'copy', callback);
 });